@@ -1,6 +1,6 @@
 /**
  * ### Модуль сборки *.js по описанию метаданных
- * &copy; Evgeniy Malyarov http://www.oknosoft.ru 2014-2017
+ * &copy; Evgeniy Malyarov http://www.oknosoft.ru 2014-2019
  * @module  metadata-prebuild
  */
 
@@ -110,7 +110,7 @@ $p.wsql.init((prm) => {
       debug('Получаем скрипт таблиц alasql');
       $p.md.create_tables((sql) => {
 
-        text = '/* eslint-disable */\nmodule.exports = function meta($p) {\n\n'
+        text = '/* eslint-disable */\nexport function meta_init($p) {\n\n'
           + '$p.wsql.alasql(\'' + sql + '\', []);\n\n'
           + '$p.md.init(' + JSON.stringify(_m) + ');\n\n'
           + text + '};';
@@ -143,21 +143,21 @@ function create_modules(_m) {
 
   const sys_nsmes = ['log', 'meta_objs', 'meta_fields', 'scheme_settings'];
   const categoties = {
-      cch: {mgr: 'ChartOfCharacteristicManager', proto: 'CatObj'},
+      cch: {mgr: 'ChartOfCharacteristicManager', proto: 'CatObj', dir: 'chartscharacteristics'},
       cacc: {mgr: 'ChartOfAccountManager', proto: 'CatObj'},
       cat: {mgr: 'CatManager', proto: 'CatObj', dir: 'catalogs'},
       bp: {mgr: 'BusinessProcessManager', proto: 'BusinessProcessObj'},
       tsk: {mgr: 'TaskManager', proto: 'TaskObj'},
-      doc: {mgr: 'DocManager', proto: 'DocObj'},
+      doc: {mgr: 'DocManager', proto: 'DocObj', dir: 'documents'},
       ireg: {mgr: 'InfoRegManager', proto: 'RegisterRow'},
       areg: {mgr: 'AccumRegManager', proto: 'RegisterRow'},
       dp: {mgr: 'DataProcessorsManager', proto: 'DataProcessorObj'},
-      rep: {mgr: 'DataProcessorsManager', proto: 'DataProcessorObj'},
+      rep: {mgr: 'DataProcessorsManager', proto: 'DataProcessorObj', dir: 'reports'},
     };
   let text = `(function(){
-  const {EnumManager,CatManager,DocManager,DataProcessorsManager,ChartOfCharacteristicManager,ChartOfAccountManager,
-    InfoRegManager,AccumRegManager,BusinessProcessManager,TaskManager,CatObj, DocObj, TabularSectionRow, DataProcessorObj,
-    RegisterRow, BusinessProcessObj, TaskObj} = $p.constructor.classes;
+  const {MetaEventEmitter,EnumManager,CatManager,DocManager,DataProcessorsManager,ChartOfCharacteristicManager,ChartOfAccountManager,
+    InfoRegManager,AccumRegManager,BusinessProcessManager,TaskManager,CatObj,DocObj,TabularSectionRow,DataProcessorObj,
+    RegisterRow,BusinessProcessObj,TaskObj} = $p.constructor.classes;
     
   const _define = Object.defineProperties;
 
@@ -193,10 +193,16 @@ function obj_constructor_text(_m, category, name, categoties) {
     f, props = '';
 
   const filename = dir && path.resolve(__dirname, `../src/metadata/${dir}/${category}_${name}.js`);
-  const extModule = dir && fs.existsSync(filename) && require(filename);
+  let extModule;
+  if(dir && fs.existsSync(filename)) {
+    try {
+      extModule = require(filename);
+    }
+    catch(err) {}
+  };
 
   const extender = extModule && extModule[fn_name] && extModule[fn_name].toString();
-  const extText = extender && extender.substring(extender.indexOf('{') + 1, extender.lastIndexOf('}') - 1);
+  const objText = extender && extender.substring(extender.indexOf('{') + 1, extender.lastIndexOf('}') - 1);
 
   const substitute = extModule && extModule.substitute && extModule.substitute.toString();
   const substituteText = substitute && substitute.substring(substitute.indexOf('{') + 3, substitute.lastIndexOf('}'));
@@ -211,32 +217,43 @@ function obj_constructor_text(_m, category, name, categoties) {
   text += '\n* @constructor \n*/\n';
   text += `class ${fn_name} extends ${proto}{\n`;
 
-  // реквизиты по метаданным
-  if (meta.fields) {
-    for (f in meta.fields) {
-      text += `get ${f}(){return this._getter('${f}')}\nset ${f}(v){this._setter('${f}',v)}\n`;
-    }
+  // если описан конструктор объекта, используем его
+  if(objText && extModule[fn_name]._replace){
+    text += objText;
   }
   else {
-    for (f in meta.dimensions) {
-      text += `get ${f}(){return this._getter('${f}')}\nset ${f}(v){this._setter('${f}',v)}\n`;
+    // реквизиты по метаданным
+    if (meta.fields) {
+      for (f in meta.fields) {
+        if(category === 'cch' && f === 'type') {
+          text += `get type(){const {type} = this._obj; return typeof type === 'object' ? type : {types: []}}
+        set type(v){this._obj.type = typeof v === 'object' ? v : {types: []}}\n`;
+        }
+        else {
+          text += `get ${f}(){return this._getter('${f}')}\nset ${f}(v){this._setter('${f}',v)}\n`;
+        }
+      }
     }
-    for (f in meta.resources) {
-      text += `get ${f}(){return this._getter('${f}')}\nset ${f}(v){this._setter('${f}',v)}\n`;
+    else {
+      for (f in meta.dimensions) {
+        text += `get ${f}(){return this._getter('${f}')}\nset ${f}(v){this._setter('${f}',v)}\n`;
+      }
+      for (f in meta.resources) {
+        text += `get ${f}(){return this._getter('${f}')}\nset ${f}(v){this._setter('${f}',v)}\n`;
+      }
+      for (f in meta.attributes) {
+        text += `get ${f}(){return this._getter('${f}')}\nset ${f}(v){this._setter('${f}',v)}\n`;
+      }
     }
-    for (f in meta.attributes) {
-      text += `get ${f}(){return this._getter('${f}')}\nset ${f}(v){this._setter('${f}',v)}\n`;
-    }
-  }
 
-  // табличные части по метаданным - устанавливаем геттер и сеттер для табличной части
-  for (const ts in meta.tabular_sections) {
-    text += `get ${ts}(){return this._getter_ts('${ts}')}\nset ${ts}(v){this._setter_ts('${ts}',v)}\n`;
-  }
+    // табличные части по метаданным - устанавливаем геттер и сеттер для табличной части
+    for (const ts in meta.tabular_sections) {
+      text += `get ${ts}(){return this._getter_ts('${ts}')}\nset ${ts}(v){this._setter_ts('${ts}',v)}\n`;
+    }
 
-  // если описан расширитель объекта, дополняем
-  if(extText){
-    text += extText;
+    if(objText){
+      text += objText;
+    }
   }
 
   text += `}\n`;
@@ -263,7 +280,7 @@ function obj_constructor_text(_m, category, name, categoties) {
 
   // если описан расширитель менеджера, дополняем
   if(managerText){
-    text += managerText.replace('extends Object', 'extends CatManager');
+    text += managerText.replace('extends Object', `extends ${mgr}`);
     text += `\n$p.${category}.create('${name}', ${managerName}, ${extModule[managerName]._freeze ? 'true' : 'false'});\n`;
   }
   else{
