@@ -189,7 +189,7 @@ class AppView extends Component {
               <MenuIcon color="inherit"/>
             </IconButton>
 
-            <Typography className={classes.title} variant="title" color="textSecondary" noWrap>{title || 'Metadata'}</Typography>
+            <Typography className={classes.title} variant="h6" color="textSecondary" noWrap>{title || 'Metadata'}</Typography>
 
             <HeaderButtons
               sync_started={sync_started}
