@@ -31,11 +31,11 @@ function PageHome(props) {
         <div className={classes.content}>
 
           <div className={classes.text}>
-            <Typography variant="display1" component="h1" color="inherit" noWrap style={{marginBottom: 24}}>Metadata hello world</Typography>
-            <Typography variant="subheading" component="h2" color="inherit" className={classes.headline}>
+            <Typography variant="h4" component="h1" color="inherit" noWrap style={{marginBottom: 24}}>Metadata hello world</Typography>
+            <Typography variant="subtitle1" component="h2" color="inherit" className={classes.headline}>
               Каркас приложения на metadata.js<br/> для демонстрации базовых возможностей фреймворка
             </Typography>
-            <Button className={classes.button} variant="raised" onClick={() => handleNavigate('/help')}>Начать</Button>
+            <Button className={classes.button} variant="contained" onClick={() => handleNavigate('/help')}>Начать</Button>
           </div>
         </div>
       </div>
